+++ .php-cs-fixer.dist.php
@@ -2,29 +2,28 @@
 
 declare(strict_types=1);
 
-$finder = (new PhpCsFixer\Finder())
-	->in(__DIR__.'/src')
-	->in(__DIR__.'/tests')
-;
+$finder = PhpCsFixer\Finder::create()
+	->in([
+		__DIR__ . '/src',
+		__DIR__ . '/tests',
+		__DIR__ . '/examples',
+	])
+	->name('*.php');
 
 return (new PhpCsFixer\Config())
+	->setRiskyAllowed(true)
+	->setIndent("\t")
 	->setRules([
 		'@PER-CS2.0' => true,
+		'@PER-CS2.0:risky' => true,
 		'declare_strict_types' => true,
+		'strict_param' => true,
+		'array_syntax' => ['syntax' => 'short'],
 		'no_unused_imports' => true,
 		'ordered_imports' => ['sort_algorithm' => 'alpha'],
-		'concat_space' => ['spacing' => 'none'],
-		'array_syntax' => ['syntax' => 'short'],
 		'single_quote' => true,
 		'trailing_comma_in_multiline' => ['elements' => ['arguments', 'arrays', 'parameters']],
-		'global_namespace_import' => [
-			'import_classes' => true,
-			'import_constants' => false,
-			'import_functions' => false,
-		],
+		'native_function_invocation' => ['include' => ['@compiler_optimized'], 'scope' => 'namespaced', 'strict' => true],
+		'concat_space' => ['spacing' => 'none'],
 	])
-	->setIndent("\t")
-	->setLineEnding("\n")
-	->setFinder($finder)
-	->setRiskyAllowed(true)
-;
+	->setFinder($finder);